Refactor ListCommand logic to have more OOP
- Convert ListCommand to an abstract class and create subclasses to handle the actual logic of listing instead
- Update ListCommandParser to reflect this change
